% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlcolor.R
\docType{package}
\name{inlcolor-package}
\alias{inlcolor}
\alias{inlcolor-package}
\title{inlcolor: Color Schemes for the USGS Idaho National Laboratory Project Office}
\description{
A collection of functions for creating color schemes. Used to support packages and scripts written by researchers at the United States Geological Survey (USGS) Idaho National Laboratory Project Office.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rconnect.usgs.gov/INLPO/inlcolor-main/}
  \item \url{https://code.usgs.gov/inl/inlcolor}
  \item Report bugs at \url{https://code.usgs.gov/inl/inlcolor/-/issues}
}

}
\author{
\strong{Maintainer}: Jason C. Fisher \email{jfisher@usgs.gov} (\href{https://orcid.org/0000-0001-9032-8912}{ORCID})

}
\keyword{internal}
