% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_color.R
\name{is_color}
\alias{is_color}
\title{Check whether a character string is a valid color specification.}
\usage{
is_color(x, null_ok = FALSE)
}
\arguments{
\item{x}{'character' vector.
color specification}

\item{null_ok}{'logical' flag.
If set to \code{TRUE}, \code{x} may also be \code{NULL}.}
}
\value{
A 'logical' vector of the same length as argument \code{x}.
}
\description{
Check whether a character string is a valid color specification.
}
\examples{
is_color(c("red", "zzz", "#FFFFFF", "#7FAF1B111"))
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
