% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.inlpal}
\alias{plot.inlpal}
\title{Plot method for color palettes}
\usage{
\method{plot}{inlpal}(x, ..., label = TRUE)
}
\arguments{
\item{x}{'inlpal' object that inherits behavior from the 'character' class.
Palette colors represented in hexadecimal format.}

\item{...}{Not used}

\item{label}{'logical' flag.
Whether to include the plot title.}
}
\value{
Invisibly returns \code{NULL}, called for side effect,
plotting a color palette in an \R graphics device.
}
\description{
Plot a sequence of shaded rectangles showing colors in the palette.
}
\examples{
get_colors(10) |> plot()

set_hinge(c(-10, 10), hinge = 0)(20) |> plot()
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
